+++ docs/examples/code_examples/add_data_to_dataset_dataset.py
@@ -1,3 +1,5 @@
+import asyncio
+
 from crawlee.storages import Dataset
 
 
@@ -7,3 +9,7 @@ async def main() -> None:
 
     # Interact with dataset directly.
     await dataset.push_data({'key': 'value'})
+
+
+if __name__ == '__main__':
+    asyncio.run(main())