+++ docs/introduction/code_examples/06_scraping.py
@@ -6,7 +6,7 @@ from crawlee.crawlers import PlaywrightCrawler, PlaywrightCrawlingContext
 async def main() -> None:
     crawler = PlaywrightCrawler(
         # Let's limit our crawls to make our tests shorter and safer.
-        max_requests_per_crawl=50,
+        max_requests_per_crawl=10,
     )
 
     @crawler.router.default_handler