+++ src/crawlee/crawlers/_playwright/_playwright_crawling_context.py
@@ -12,7 +12,7 @@ if TYPE_CHECKING:
 
     from playwright.async_api import Response
 
-    from crawlee._types import EnqueueLinksFunction
+    from crawlee._types import EnqueueLinksFunction, ExtractLinksFunction
 
 
 @dataclass(frozen=True)
@@ -28,6 +28,9 @@ class PlaywrightCrawlingContext(PlaywrightPreNavCrawlingContext):
 
     enqueue_links: EnqueueLinksFunction
     """The Playwright `EnqueueLinksFunction` implementation."""
+
+    extract_links: ExtractLinksFunction
+    """The Playwright `ExtractLinksFunction` implementation."""
 
     infinite_scroll: Callable[[], Awaitable[None]]
     """A function to perform infinite scrolling on the page. This scrolls to the bottom, triggering