+++ tests/unit/crawlers/_basic/test_basic_crawler.py
@@ -14,7 +14,6 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal, cast
 from unittest.mock import AsyncMock, Mock, call
 
-import httpx
 import pytest
 
 from crawlee import ConcurrencySettings, Glob, service_locator
@@ -22,18 +21,18 @@ from crawlee._request import Request
 from crawlee._types import BasicCrawlingContext, EnqueueLinksKwargs, HttpHeaders
 from crawlee.configuration import Configuration
 from crawlee.crawlers import BasicCrawler
-from crawlee.errors import SessionError, UserDefinedErrorHandlerError
+from crawlee.errors import RequestCollisionError, SessionError, UserDefinedErrorHandlerError
 from crawlee.events._local_event_manager import LocalEventManager
 from crawlee.request_loaders import RequestList, RequestManagerTandem
-from crawlee.sessions import SessionPool
+from crawlee.sessions import Session, SessionPool
 from crawlee.statistics import FinalStatistics
 from crawlee.storage_clients import MemoryStorageClient
 from crawlee.storages import Dataset, KeyValueStore, RequestQueue
 
 if TYPE_CHECKING:
-    from collections.abc import Sequence
+    from collections.abc import Callable, Sequence
 
-    import respx
+    from yarl import URL
 
     from crawlee._types import JsonSerializable
     from crawlee.storage_clients._memory import DatasetClient
@@ -300,35 +299,29 @@ async def test_handles_error_in_failed_request_handler() -> None:
         await crawler.run(['http://a.com/', 'http://b.com/', 'http://c.com/'])
 
 
-async def test_send_request_works(respx_mock: respx.MockRouter) -> None:
-    respx_mock.get('http://b.com/', name='test_endpoint').return_value = httpx.Response(
-        status_code=200, json={'hello': 'world'}
-    )
+async def test_send_request_works(server_url: URL) -> None:
+    response_data: dict[str, Any] = {}
 
-    response_body: Any = None
-    response_headers: HttpHeaders | None = None
-
-    crawler = BasicCrawler(
-        max_request_retries=3,
-    )
+    crawler = BasicCrawler(max_request_retries=3)
 
     @crawler.router.default_handler
     async def handler(context: BasicCrawlingContext) -> None:
-        nonlocal response_body, response_headers
+        response = await context.send_request(str(server_url))
 
-        response = await context.send_request('http://b.com/')
-        response_body = response.read()
-        response_headers = response.headers
+        response_data['body'] = response.read()
+        response_data['headers'] = response.headers
 
     await crawler.run(['http://a.com/', 'http://b.com/', 'http://c.com/'])
-    assert respx_mock['test_endpoint'].called
 
-    assert json.loads(response_body) == {'hello': 'world'}
+    response_body = response_data.get('body')
+    assert response_body is not None
+    assert b'Hello, world!' in response_body
 
+    response_headers = response_data.get('headers')
     assert response_headers is not None
     content_type = response_headers.get('content-type')
     assert content_type is not None
-    assert content_type.endswith('/json')
+    assert content_type == 'text/html; charset=utf-8'
 
 
 @dataclass
@@ -1217,3 +1210,84 @@ async def test_session_retire_in_user_handler(*, retire: bool) -> None:
         assert sessions[1] != sessions[0]
     else:
         assert sessions[1] == sessions[0]
+
+
+async def test_bound_session_to_request() -> None:
+    async with SessionPool() as session_pool:
+        check_session: Session = await session_pool.get_session()
+        used_sessions = list[str]()
+        crawler = BasicCrawler(session_pool=session_pool)
+
+        @crawler.router.default_handler
+        async def handler(context: BasicCrawlingContext) -> None:
+            if context.session:
+                used_sessions.append(context.session.id)
+
+        requests = [
+            Request.from_url('http://a.com/', session_id=check_session.id, always_enqueue=True) for _ in range(10)
+        ]
+
+        await crawler.run(requests)
+
+        assert len(used_sessions) == 10
+        assert set(used_sessions) == {check_session.id}
+
+
+async def test_bound_sessions_to_same_request() -> None:
+    # Use a custom function to avoid errors due to random Session retrieval
+    def create_session_function() -> Callable[[], Session]:
+        counter = -1
+
+        def create_session() -> Session:
+            nonlocal counter
+            counter += 1
+            return Session(id=str(counter))
+
+        return create_session
+
+    check_sessions = [str(session_id) for session_id in range(10)]
+    used_sessions = list[str]()
+    crawler = BasicCrawler(session_pool=SessionPool(create_session_function=create_session_function()))
+
+    @crawler.router.default_handler
+    async def handler(context: BasicCrawlingContext) -> None:
+        if context.session:
+            used_sessions.append(context.session.id)
+
+    requests = [
+        Request.from_url('http://a.com/', session_id=str(session_id), use_extended_unique_key=True)
+        for session_id in range(10)
+    ]
+
+    await crawler.run(requests)
+
+    assert len(used_sessions) == 10
+    assert set(used_sessions) == set(check_sessions)
+
+
+async def test_error_bound_session_to_request() -> None:
+    crawler = BasicCrawler(request_handler=AsyncMock())
+
+    requests = [Request.from_url('http://a.com/', session_id='1', always_enqueue=True) for _ in range(10)]
+
+    stats = await crawler.run(requests)
+
+    assert stats.requests_total == 10
+    assert stats.requests_failed == 10
+    assert stats.retry_histogram == [10]
+
+
+async def test_handle_error_bound_session_to_request() -> None:
+    error_handler_mock = AsyncMock()
+    crawler = BasicCrawler(request_handler=AsyncMock())
+
+    @crawler.failed_request_handler
+    async def error_req_hook(context: BasicCrawlingContext, error: Exception) -> None:
+        if isinstance(error, RequestCollisionError):
+            await error_handler_mock(context, error)
+
+    requests = [Request.from_url('http://a.com/', session_id='1')]
+
+    await crawler.run(requests)
+
+    assert error_handler_mock.call_count == 1