+++ website/src/pages/index.js
@@ -14,7 +14,10 @@ import HomepageCtaSection from '../components/Homepage/HomepageCtaSection';
 import HomepageHeroSection from '../components/Homepage/HomepageHeroSection';
 import LanguageInfoWidget from '../components/Homepage/LanguageInfoWidget';
 import RiverSection from '../components/Homepage/RiverSection';
+import RunnableCodeBlock from '../components/RunnableCodeBlock';
 import ThreeCardsWithIcon from '../components/Homepage/ThreeCardsWithIcon';
+
+import HomePageExample from '!!raw-loader!roa-loader!./home_page_example.py';
 
 function GetStartedSection() {
     return (
@@ -28,57 +31,14 @@ function GetStartedSection() {
     );
 }
 
-const example = `import asyncio
-
-from crawlee.crawlers import PlaywrightCrawler, PlaywrightCrawlingContext
-
-
-async def main() -> None:
-    crawler = PlaywrightCrawler(
-        max_requests_per_crawl=5,  # Limit the crawl to 5 requests at most.
-        headless=False,  # Show the browser window.
-        browser_type='firefox',  # Use the Firefox browser.
-    )
-
-    # Define the default request handler, which will be called for every request.
-    @crawler.router.default_handler
-    async def request_handler(context: PlaywrightCrawlingContext) -> None:
-        context.log.info(f'Processing {context.request.url} ...')
-
-        # Extract and enqueue all links found on the page.
-        await context.enqueue_links()
-
-        # Extract data from the page using Playwright API.
-        data = {
-            'url': context.request.url,
-            'title': await context.page.title(),
-            'content': (await context.page.content())[:100],
-        }
-
-        # Push the extracted data to the default dataset.
-        await context.push_data(data)
-
-    # Run the crawler with the initial list of URLs.
-    await crawler.run(['https://crawlee.dev'])
-
-    # Export the entire dataset to a JSON file.
-    await crawler.export_data('results.json')
-
-    # Or work with the data directly.
-    data = await crawler.get_data()
-    crawler.log.info(f'Extracted data: {data.items}')
-
-
-if __name__ == '__main__':
-    asyncio.run(main())
-`;
-
 function CodeExampleSection() {
     return (
         <section className={styles.codeExampleSection}>
             <div className={styles.decorativeRow} />
             <div className={styles.codeBlockContainer}>
-                <CodeBlock language="python">{example}</CodeBlock>
+                <RunnableCodeBlock className="language-python" language="python">
+                    {HomePageExample}
+                </RunnableCodeBlock>
             </div>
             <div className={styles.dashedSeparator} />
             <div className={styles.decorativeRow} />
@@ -124,8 +84,8 @@ function BenefitsSection() {
                     <ThemedImage
                         alt="Work with your favorite tools"
                         sources={{
-                            light: 'img/favorite-tools-light.webp',
-                            dark: 'img/favorite-tools-dark.webp',
+                            light: '/python/img/favorite-tools-light.webp',
+                            dark: '/python/img/favorite-tools-dark.webp',
                         }}
                     />
                 }
@@ -156,8 +116,8 @@ function OtherFeaturesSection() {
                     <Link className={styles.cardWithImage} to="/docs/guides/scaling-crawlers">
                         <ThemedImage
                             sources={{
-                                light: 'img/auto-scaling-light.webp',
-                                dark: 'img/auto-scaling-dark.webp',
+                                light: '/python/img/auto-scaling-light.webp',
+                                dark: '/python/img/auto-scaling-dark.webp',
                             }}
                             alt=""
                         />
@@ -176,8 +136,8 @@ function OtherFeaturesSection() {
                     <Link className={styles.cardWithImage} to="/docs/guides/proxy-management">
                         <ThemedImage
                             sources={{
-                                light: 'img/smart-proxy-light.webp',
-                                dark: 'img/smart-proxy-dark.webp',
+                                light: '/python/img/smart-proxy-light.webp',
+                                dark: '/python/img/smart-proxy-dark.webp',
                             }}
                             alt=""
                         />
@@ -200,8 +160,8 @@ function OtherFeaturesSection() {
                             icon: (
                                 <ThemedImage
                                     sources={{
-                                        light: 'img/queue-light-icon.svg',
-                                        dark: 'img/queue-dark-icon.svg',
+                                        light: '/python/img/queue-light-icon.svg',
+                                        dark: '/python/img/queue-dark-icon.svg',
                                     }}
                                     alt=""
                                 />
@@ -215,8 +175,8 @@ function OtherFeaturesSection() {
                             icon: (
                                 <ThemedImage
                                     sources={{
-                                        light: 'img/scraping-utils-light-icon.svg',
-                                        dark: 'img/scraping-utils-dark-icon.svg',
+                                        light: '/python/img/scraping-utils-light-icon.svg',
+                                        dark: '/python/img/scraping-utils-dark-icon.svg',
                                     }}
                                     alt=""
                                 />
@@ -231,8 +191,8 @@ function OtherFeaturesSection() {
                             icon: (
                                 <ThemedImage
                                     sources={{
-                                        light: 'img/routing-light-icon.svg',
-                                        dark: 'img/routing-dark-icon.svg',
+                                        light: '/python/img/routing-light-icon.svg',
+                                        dark: '/python/img/routing-dark-icon.svg',
                                     }}
                                     alt=""
                                 />
@@ -318,8 +278,8 @@ function BuildFastScrapersSection() {
                         icon: (
                             <ThemedImage
                                 sources={{
-                                    light: 'img/zero-setup-light-icon.svg',
-                                    dark: 'img/zero-setup-dark-icon.svg',
+                                    light: '/python/img/zero-setup-light-icon.svg',
+                                    dark: '/python/img/zero-setup-dark-icon.svg',
                                 }}
                                 alt=""
                             />
@@ -336,8 +296,8 @@ function BuildFastScrapersSection() {
                         icon: (
                             <ThemedImage
                                 sources={{
-                                    light: 'img/defaults-light-icon.svg',
-                                    dark: 'img/defaults-dark-icon.svg',
+                                    light: '/python/img/defaults-light-icon.svg',
+                                    dark: '/python/img/defaults-dark-icon.svg',
                                 }}
                                 alt=""
                             />
@@ -354,8 +314,8 @@ function BuildFastScrapersSection() {
                         icon: (
                             <ThemedImage
                                 sources={{
-                                    light: 'img/community-light-icon.svg',
-                                    dark: 'img/community-dark-icon.svg',
+                                    light: '/python/img/community-light-icon.svg',
+                                    dark: '/python/img/community-dark-icon.svg',
                                 }}
                                 alt=""
                             />